+++ sandpack-react/src/components/CodeEditor/CodeMirror.tsx
@@ -165,6 +165,16 @@ export const CodeMirror = React.forwardRef<CodeMirrorRef, CodeMirrorProps>(
       code,
     });
 
+    // decorators need to be sorted by `line`, otherwise it will throw error
+    // see https://github.com/codesandbox/sandpack/issues/383
+    const sortedDecorators = React.useMemo(
+      () =>
+        decorators
+          ? decorators.sort((d1, d2) => d1.line - d2.line)
+          : decorators,
+      [decorators]
+    );
+
     React.useEffect(() => {
       if (!wrapper.current || !shouldInitEditor) return;
 
@@ -223,14 +233,15 @@ export const CodeMirror = React.forwardRef<CodeMirrorRef, CodeMirrorProps>(
         ];
 
         if (readOnly) {
+          extensionList.push(EditorState.readOnly.of(true));
           extensionList.push(EditorView.editable.of(false));
         } else {
           extensionList.push(bracketMatching());
           extensionList.push(highlightActiveLine());
         }
 
-        if (decorators) {
-          extensionList.push(highlightDecorators(decorators));
+        if (sortedDecorators) {
+          extensionList.push(highlightDecorators(sortedDecorators));
         }
 
         if (wrapContent) {
@@ -293,7 +304,13 @@ export const CodeMirror = React.forwardRef<CodeMirrorRef, CodeMirrorProps>(
 
       // TODO: Would be nice to reconfigure the editor when these change, instead of recreating with all the extensions from scratch
       // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [shouldInitEditor, showLineNumbers, wrapContent, themeId, decorators]);
+    }, [
+      shouldInitEditor,
+      showLineNumbers,
+      wrapContent,
+      themeId,
+      sortedDecorators,
+    ]);
 
     React.useEffect(() => {
       // When the user clicks on a tab button on a larger screen