+++ sandpack-react/src/components/CodeViewer/CodeViewer.stories.tsx
@@ -12,50 +12,26 @@ export default {
   component: SandpackCodeViewer,
 };
 
-export const Component: Story<CodeViewerProps> = (args) => (
-  <SandpackProvider
-    customSetup={{
-      entry: "/index.js",
-    }}
-    files={{
-      "/index.js": {
-        code: `const people = [{
-  id: 0,
-  name: 'Creola Katherine Johnson',
-  profession: 'mathematician',
-}, {
-  id: 1,
-  name: 'Mario José Molina-Pasquel Henríquez',
-  profession: 'chemist',
-}];
-
-export default function List() {
-  const listItems = people.map(person =>
-    <li>{person}</li>
-  );
-  return <ul>{listItems}</ul>;
-}`,
-      },
-    }}
-  >
+export const Component: Story<CodeViewerProps> = () => (
+  <SandpackProvider>
     <SandpackThemeProvider>
-      <SandpackCodeViewer showLineNumbers {...args} />
+      <SandpackCodeViewer showTabs={false} />
     </SandpackThemeProvider>
   </SandpackProvider>
 );
 
-export const ReactCode: React.FC = () => (
-  <SandpackProvider template="react">
+export const ShowTabs: Story<CodeViewerProps> = () => (
+  <SandpackProvider>
     <SandpackThemeProvider>
-      <SandpackCodeViewer />
+      <SandpackCodeViewer showTabs />
     </SandpackThemeProvider>
   </SandpackProvider>
 );
 
-export const VueCode: React.FC = () => (
-  <SandpackProvider template="vue">
-    <SandpackThemeProvider theme="dark">
-      <SandpackCodeViewer />
+export const showLineNumbers: Story<CodeViewerProps> = () => (
+  <SandpackProvider>
+    <SandpackThemeProvider>
+      <SandpackCodeViewer showTabs={false} showLineNumbers />
     </SandpackThemeProvider>
   </SandpackProvider>
 );
@@ -126,17 +102,17 @@ export default function List() {
             { className: "highlight", line: 9 },
             {
               className: "widget",
-              elementAttributes: { "data-id": "1" },
-              line: 12,
-              startColumn: 26,
-              endColumn: 38,
-            },
-            {
-              className: "widget",
               elementAttributes: { "data-id": "2" },
               line: 13,
               startColumn: 8,
               endColumn: 17,
+            },
+            {
+              className: "widget",
+              elementAttributes: { "data-id": "1" },
+              line: 12,
+              startColumn: 26,
+              endColumn: 38,
             },
           ]}
           showLineNumbers={false}